+++ src/dipdup/dipdup.py
@@ -599,6 +599,7 @@ class DipDup:
     ) -> None:
         """Create new or update existing dipdup project"""
         from dipdup.codegen.evm import EvmCodeGenerator
+        from dipdup.codegen.starknet import StarknetCodeGenerator
         from dipdup.codegen.tezos import TezosCodeGenerator
 
         await self._create_datasources()
@@ -612,6 +613,7 @@ class DipDup:
             codegen_classes: tuple[type[CodeGenerator], ...] = (
                 TezosCodeGenerator,
                 EvmCodeGenerator,
+                StarknetCodeGenerator,
             )
             for codegen_cls in codegen_classes:
                 codegen = codegen_cls(