+++ src/dipdup/models/starknet.py
@@ -1,9 +1,23 @@
 from abc import ABC
 from dataclasses import dataclass
 from typing import Any
+from typing import Generic
+from typing import Literal
 from typing import Self
+from typing import TypeVar
+
+from pydantic import BaseModel
 
 from dipdup.fetcher import HasLevel
+from dipdup.subscriptions import Subscription
+
+
+@dataclass(frozen=True)
+class StarknetSubscription(Subscription):
+    name: Literal['starknet'] = 'starknet'
+
+    def get_params(self) -> list[Any]:
+        return [self.name]
 
 
 @dataclass(frozen=True)
@@ -80,3 +94,12 @@ class StarknetEventData(HasLevel, ABC):
             keys=tuple(event_json['keys']),
             data=tuple(event_json['data']),
         )
+
+
+PayloadT = TypeVar('PayloadT', bound=BaseModel)
+
+
+@dataclass(frozen=True)
+class StarknetEvent(Generic[PayloadT]):
+    data: StarknetEventData
+    payload: PayloadT